servidor/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic-settings
motor
pymongo
passlib[bcrypt]
python-jose[cryptography]